Add unit tests for the stringifyBody
Also fixed a wrong character code value found in testing.

Old implementations of stringifyBody and sortObject are now included in the rsa.test.js but can be
removed when the new ones have been validated. There is a way to make the old
implementation failures to pass the test suite, but they are disabled
for now to show the pitfalls of the old implementation.